Moved File, Added File
Moved musicinterface out of music directory at Emberling's request. Added the README back to the music directory.
